add job listener, log entry step and fix the query by adding order by which will resolve the same member picked in multiple threads issue

+++ src/main/java/com/guptaji/springBatchPartitionerDemo/controller/DataCreator.java
@@ -24,7 +24,7 @@ public class DataCreator {
   public ResponseEntity<?> createData(
       @PathVariable("name") String name,
       @PathVariable("clgName") String clgName,
-      @PathVariable("count") int count) {
+      @PathVariable("count") long count) {
     studentOneService.createData(name, clgName, count);
     LOG.info("Created data");
     return ResponseEntity.status(HttpStatus.OK).body("Created the data");
