+++ src/main/java/com/guptaji/springBatchPartitionerDemo/entity/StudentTwo.java
@@ -6,16 +6,16 @@ import jakarta.persistence.Id;
 @Entity
 public class StudentTwo {
 
-  @Id private int id;
+  @Id private Long id;
 
   private String name;
   private String collegeName;
 
-  public int getId() {
+  public Long getId() {
     return id;
   }
 
-  public void setId(int id) {
+  public void setId(Long id) {
     this.id = id;
   }
