+++ src/main/java/com/guptaji/springBatchPartitionerDemo/repositories/StudentTwoRepo.java
@@ -6,4 +6,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StudentTwoRepo extends JpaRepository<StudentTwo, Integer> {}
+public interface StudentTwoRepo extends JpaRepository<StudentTwo, Long> {}
